launcher pagination, mouse fix, refactor key handling. editor. rework boot sequence.
